Respect Do Not Track (DNT) headers by default
The Do Not Track (DNT) header is the proposed HTTP header that requests
that a web application disable its tracking of an individual user.

This changes the default to respect the DNT header and no longer injects
any tracker if the DNT header is set to 1 (user opt-out).
This behaviour can be overwritten by setting the
DO_NOT_RESPECT_THE_USERS_CHOICE_TO_OPT_OUT option on handlers that
should still be injected.